Added Music Beginnings
- Added the play command, which for an uknown reason dosen't play the music

- Added the queue command which sends the current queued songs to chat

- Standardized the "..ID" captitilization

- Added the skip command which will in future skip songs

- Added a clause for no error message being provided in the errorHandler.js file

- Added a clause to check that a message still exists before deleting it in the send() function in message.js

- Added the clientAction emitter & handler*

- Added the musicManager.js file

? Key
* - May be removed in future updates

+++ src/services/events/message.js
@@ -91,7 +91,7 @@ module.exports = message => {
         message.channel.send(text).then(message => {
             if(timeout == 0) return;
             setTimeout(() => {
-                message.delete();
+                if(message) message.delete();
             }, timeout); 
         }); 
     }
